pyproject.toml
[tool.poetry]
name = "opde"
version = "0.1.0"
description = ""
authors = ["ElonH <elonhhuang@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
PyGithub = "^1.51"
ply = "^3.11"
networkx = "^2.4"
matplotlib = "^3.2.2"
gitpython = "^3.1.7"
Click = "^7.1.2"
tinydb = "^4.1.1"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
notebook = "^6.0.3"
rope = "^0.17.0"
pylint = "^2.5.3"
autopep8 = "^1.5.3"
cmake_format = "^0.6.10"
python-language-server = {extras = ["all"], version = "^0.34.1"}
jupyterlab = "^2.2.0"
jupyter-lsp = "^0.9.0"
pyyaml = "^5.3.1"

[build-system]
requires = ["poetry>=0.12", "cmake>=3.2.0"]
build-backend = "poetry.masonry.api"
